+++ unpack/BCSAR.py
@@ -1,305 +1,322 @@
 # -*- coding:utf-8 -*-
-from util import error, ENDIANS
-from util.funcops import getsup, byterepr
+import os
+from util import error, warning, ENDIANS
+from util.funcops import byterepr
 from util.fileops import *
 import util.rawutil as rawutil
 
-CSAR_HEADER_STRUCT = '4s2H13I'
-STRG_HEADER_STRUCT = '4s5I'
-INFO_HEADER_STRUCT = '4s17I'
-INFO_SECTION_STRUCT = 'I/0[2I]$'
+CSAR_HEADER = '4s2H2I2H /p2[2H2I]'
+STRG_HEADER = '4sI 2[2HI]'
+STRG_STR_OFFSETTABLE = 'I /p1[2H2I]'
+INFO_HEADER = '4sI 8[2HI]'
+INFO_SECTION = 'I/0[2I]'
 
-BANK_ENTRY_STRUCT = '6I'
-PLAYER_ENTRY_STRUCT = '4I'
-WAR_ENTRY_STRUCT = '3I'
-GROUP_ENTRY_STRUCT = '3I'
+# INFO tables
+class INFOTable (rawutil.TypeReader):
+	byteorder = None
+	bcsar = None
+	def __init__(self, offset):
+		self.offset = offset
 
 
-class FileNode (object):
+class TableEntry (object):
+	def __init__(self, type=None, **kwargs):
+		self.type = type
+		self.__dict__.update(kwargs)
+	
+	def __repr__(self):
+		return '%s entry: %s' % (self.type, ', '.join(['%s = %s' % (el, self.__dict__[el]) for el in self.__dict__]))
+
+
+class AudioTable (INFOTable):
+	id = 0x2100
+	offsetid = 0x2200
+	def extract(self, data, offsettbl):
+		self.entries = []
+		for offsetid, offset in offsettbl:
+			fileid = self.unpack_from('I', data, offset)[0]
+			id = self.unpack_from('H', data, offset + 12)[0]
+			nameid = self.unpack_from('I', data, offset + 24)[0]
+			name = self.bcsar.strg.strings[nameid]
+			if id == 0x2201:  #External BCSTM
+				entry = TableEntry('BCSTM', nameid=nameid, name=name, fileid=fileid)
+			elif id == 0x2202:  #WSD
+				entry = TableEntry('BCWSD', nameid=nameid, name=name, fileid=fileid)
+			elif id == 0x2203:  #SEQ
+				bank = self.unpack_from('H', data, offset + 92)
+				entry = TableEntry('BCSEQ', nameid=nameid, name=name, fileid=fileid, bank=bank)
+			self.entries.append(entry)
+			
+
+class BankTable (INFOTable):
+	id = 0x2101
+	offsetid = 0x2206
+	def extract(self, data, offsettbl):
+		self.entries = []
+		for offsetid, offset in offsettbl:
+			fileid, unk1, unk2, unk3, nameid, unk4 = self.unpack_from('6I', data, offset)
+			name = self.bcsar.strg.strings[nameid]
+			entry = TableEntry('BCBNK', fileid=fileid, nameid=nameid, name=name)
+			self.entries.append(entry)
+
+class PlayerTable (INFOTable):
+	id = 0x2102
+	def extract(self, data, offsettbl):
+		pass
+
+class WaveArcTable (INFOTable):
+	id = 0x2103
+	offsetid = 0x2207
+	def extract(self, data, offsettbl):
+		self.entries = []
+		for i, (offsetid, offset) in enumerate(offsettbl):
+			if i < len(offsettbl) - 1:
+				datalen = offsettbl[i + 1][1] - offset
+			else:
+				datalen = self.info.grouptbl.offset - offset
+			if datalen > 12:
+				fileid, unk1, unk2, nameid = self.unpack_from('4I', data, offset)
+				name = self.bcsar.strg.strings[nameid]
+			else:
+				fileid = self.unpack_from('3I', data, offset)[0]
+				name = 'WAR_%08x' % i
+				nameid = 0xffffffff
+			entry = TableEntry('BCWAR', fileid=fileid, name=name, nameid=nameid)
+			self.entries.append(entry)
+
+class SetTable (INFOTable):  #possible error
+	id = 0x2104
+	offsetid = 0x2204
+	def extract(self, data, offsettbl):
+		self.entries = []
+		for offsetid, offset in offsettbl:
+			if offset == 0xffffffff:
+				continue
+			header = self.unpack_from('I', data, offset)
+			if header == 0xffffffff:
+				continue
+			nameid, unk, fileid = self.unpack_from('3I', data, offset + 28)
+			name = self.bcsar.strg.strings[nameid]
+			entry = TableEntry('Set', fileid=fileid, nameid=nameid, name=name)
+			self.entries.append(entry)
+
+class GroupTable (INFOTable):
+	id = 0x2105
+	offsetid = 0x2208
+	def extract(self, data, offsettbl):
+		self.entries = []
+		for offsetid, offset in offsettbl:
+			fileid, unk, nameid = self.unpack_from('3I', data, offset)
+			name = self.bcsar.strg.strings[nameid]
+			entry = TableEntry('BCGRP', fileid=fileid, nameid=nameid, name=name)
+			self.entries.append(entry)
+
+class FileTable (INFOTable):
+	id = 0x2106
+	offsetid = 0x220a
+	endid = 0x220b
+	def extract(self, data, offsettbl):
+		self.entries = []
+		for offsetid, offset in offsettbl:
+			id = self.unpack_from('2H2I', data, offset)[0]
+			entry = TableEntry('File', internal=False, valid=True)
+			if id == 0x220d:  #External
+				entry.internal = False
+				entry.path = self.unpack_from('n', data)[0]
+			elif id == 0x220c:  #Internal
+				entry.internal = True
+				header, padding, offset, size = self.unpack_from('2H2I', data)
+				if header != 0xffff and offset != 0xffffffff:
+					entry.header = header
+					entry.offset = offset
+					entry.size = size
+				else:
+					entry.valid = False
+			else:
+				entry.valid = False
+			self.entries.append(entry)
+
+
+#BCSAR sections
+
+class BCSARSection (rawutil.TypeReader):
+	bcsar = None
+	byteorder = None
+
+class STRG (BCSARSection):
+	id = 0x2000
+	type = 'STRG'
 	def __init__(self):
-		self.name = ''
-		self.filename = None
-		self.internal = False
-		self.type = 'Undefined'
-		self.fileid = None
-		self.bank = None
-		self.set = None
-		self.war = None
-		self.group = None
+		self.offset = self.size = None
+		self.strings = []
+	
+	def extract(self, data):
+		header = self.unpack_from(STRG_HEADER, data, self.offset)
+		if header[0] != b'STRG':
+			error('Invalid magic %s, expected STRG' % byterepr(header[0]), 301)
+		#sectionlen = header[1]
+		for i, el in enumerate(header[2]):
+			id, padding, offset = el
+			if id == 0x2400:  #STRinGs table offset
+				self.strtbl_offset = offset + 8 + (8 * i)
+			elif id == 0x2401:  #Another offset table
+				self.othertbl_offset = offset + 8 + (8 * i)
+			else:
+				error('Invalid section ID in STRG 0x%04x' % id, 303)
+		strtbl_offsettbl = self.unpack_from(STRG_STR_OFFSETTABLE, data, self.strtbl_offset + self.offset)
+		self.strings_count = strtbl_offsettbl[0]
+		for entry in strtbl_offsettbl[1]:
+			id, padding, offset, size = entry
+			if id != 0x1f01:
+				error('Invalid entry ID in strings table offsets 0x%04x, expected 0x1f01' % id, 303)
+			else:
+				self.strings.append(self.unpack_from('n', data, offset + self.strtbl_offset + 0x40)[0].decode('utf-8'))
+		
 
-
-class BankNode (object):
+class INFO (BCSARSection):
+	id = 0x2001
+	type = 'INFO'
 	def __init__(self):
-		self.fileid = None
-		self.name = ''
+		self.offset = self.size = None
+	
+	def extract(self, data):
+		header = self.unpack_from(INFO_HEADER, data, self.offset)
+		if header[0] != b'INFO':
+			error('Invalid magic %s, expected INFO' % header[0], 301)
+		#sectionlen = header[1]
+		for i, el in enumerate(header[2]):
+			id, padding, offset = el
+			if id == AudioTable.id:
+				self.audiotbl = AudioTable(offset + self.offset + 8)
+			elif id == BankTable.id:
+				self.banktbl = BankTable(offset + self.offset + 8)
+			elif id == PlayerTable.id:
+				self.playertbl = PlayerTable(offset + self.offset + 8)
+			elif id == WaveArcTable.id:
+				self.warctbl = WaveArcTable(offset + self.offset + 8)
+			elif id == SetTable.id:
+				self.settbl = SetTable(offset + self.offset + 8)
+			elif id == GroupTable.id:
+				self.grouptbl = GroupTable(offset + self.offset + 8)
+			elif id == FileTable.id:
+				self.filetbl = FileTable(offset + self.offset + 8)
+			elif id == FileTable.endid:
+				self.filetbl.end = offset
+		self.readTable(data, self.filetbl)
+		self.readTable(data, self.warctbl)
+		self.readTable(data, self.banktbl)
+		self.readTable(data, self.audiotbl)
+		self.readTable(data, self.settbl)
+		self.readTable(data, self.grouptbl)
+	
+	def readOffsetTable(self, data, offset):
+		entrycount, offsettbl = self.unpack_from(INFO_SECTION, data, offset)
+		offsettbl = [(el[0], el[1] + offset) for el in offsettbl]
+		return offsettbl
+	
+	def readTable(self, data, obj):
+		offsettbl = self.readOffsetTable(data, obj.offset)
+		obj.extract(data, offsettbl)
+		
 
-
-class AudioNode (object):
+class FILE (BCSARSection):
+	id = 0x2002
+	type = 'FILE'
 	def __init__(self):
-		self.fileid = None
-		self.name = ''
-
-
-class SetNode (object):
-	def __init__(self):
-		self.fileid = None
-		self.name = ''
-
-
-class WARNode (object):
-	def __init__(self):
-		self.fileid = None
-		self.name = ''
-
-
-class GroupNode (object):
-	def __init__(self):
-		self.fileid = None
-		self.name = ''
-
+		self.offset = self.size = None
+	
+	def extract(self, data):
+		linkout = ''
+		info = self.bcsar.info
+		self.files = info.filetbl.entries
+		linkout += self.processTable(data, info.warctbl)
+		linkout += self.processTable(data, info.audiotbl)
+		linkout += self.processTable(data, info.banktbl)
+		linkout += self.processTable(data, info.grouptbl)
+		write(linkout, self.bcsar.outdir + 'ExternalLinkedFiles.txt')
+	
+	def processTable(self, data, table):
+		linked = []
+		for entry in table.entries:
+			fileentry = self.files[entry.fileid]
+			if fileentry.valid:
+				if fileentry.internal:
+					data.seek(fileentry.offset + self.offset + 8)
+					bwrite(data.read(fileentry.size), self.makepath(entry))
+				else:
+					linked.append((entry.name, fileentry.path))
+			else:
+				warning('Invalid entry %d' % entry.fileid, 902)
+				continue
+		linkout = ''
+		for entry in linked:
+			linkout += '%s: %s\n' % entry
+		paf
+		return linkout
+			
+	
+	def makepath(self, entry):
+		pathes = {
+			'BCWAR': self.bcsar.warcdir,
+			'BCWSD': self.bcsar.wsddir,
+			'BCBNK': self.bcsar.bankdir,
+			'BCGRP': self.bcsar.groupdir,
+			'BCSEQ': self.bcsar.seqdir,
+		}
+		filename = path(pathes[entry.type], '%s.%s' % (entry.name, entry.type))
+		return filename
 
 class extractBCSAR (rawutil.TypeReader):
-	def __init__(self, filename, data, verbose, opts={}):
-		self.make_outdirs(filename)
-		self.verbose = verbose
-		self.byteorder = ENDIANS[rawutil.unpack_from('>H', data, 4)[0]]
-		self.readheader(data)
-		self.readSTRG()
-		self.readINFO()
-
-	def make_outdirs(self, filename):
+	def __init__(self, filename, file, verbose, opts={}):
 		self.outdir = make_outdir(filename)
-		self.wardir = makedirs(self.outdir + 'WaveArchives')
+		self.make_outdirs()
+		self.verbose = verbose
+		self.data = file
+		self.data.seek(0)
+		INFOTable.bcsar = self
+		BCSARSection.bcsar = self
+		self.strg = STRG()
+		self.info = INFO()
+		self.file = FILE()
+		INFOTable.info = self.info
+		self.readheader()
+		self.strg.extract(self.data)
+		self.info.extract(self.data)
+	
+	def make_outdirs(self):
+		self.warcdir = makedirs(self.outdir + 'WaveArchives')
 		self.bankdir = makedirs(self.outdir + 'Banks')
 		self.groupdir = makedirs(self.outdir + 'Groups')
 		self.seqdir = makedirs(self.outdir + 'Sequences')
 		self.wsddir = makedirs(self.outdir + 'WaveSounds')
-
-	def readheader(self, data):
-		hdata = self.unpack_from(CSAR_HEADER_STRUCT, data, 0)
-		if hdata[0] != b'CSAR':
-			error('Invalid magic %s, expected CSAR' % byterepr(hdata[0]), 301)
-		#bom = hdata[1]
-		#headerlen = hdata[2]
-		self.version = hdata[3]
-		#filelen = hdata[4]
-		#seccount = hdata[5]
-		i = 6
-		while i < 14:
-			id, offset, length = hdata[i:i + 3]
-			i += 3
-			secdata = data[offset:offset + length]
-			if id == 0x2000:
-				self.strg = secdata
-			elif id == 0x2001:
-				self.info = secdata
-			elif id == 0x2002:
-				self.file = secdata
-				self.fileoffset = offset
+	
+	def readheader(self):
+		self.byteorder = ENDIANS[self.unpack_from('>H', self.data, 4)[0]]
+		INFOTable.byteorder = self.byteorder
+		BCSARSection.byteorder = self.byteorder
+		header = self.unpack_from(CSAR_HEADER, self.data, 0)
+		magic = header[0]
+		if magic != b'CSAR':
+			error('Invalid magic %s, expected CSAR' % byterepr(magic), 301)
+		#bom = header[1]
+		#headerlen = header[2]
+		#version = header[3]  #?
+		#filelen = header[4]
+		seccount = header[5]
+		#padding = header[6]  #?
+		for sec in header[7]:
+			id, padding, offset, size = sec
+			if id == STRG.id:  #STRG
+				self.strg.offset = offset
+				self.strg.size = size
+			elif id == INFO.id:  #INFO
+				self.info.offset = offset
+				self.info.size = size
+			elif id == FILE.id:  #FILE
+				self.file.offset = offset
+				self.file.size = size
 			else:
-				error('Unsupported partition 0x%04x' % id, 303)
-
-		#bwrite(self.strg, 'STRG')
-		#bwrite(self.info, 'INFO')
-		#bwrite(self.file, 'FILE')
-
-	def readSTRG(self):
-		hdata, ptr = self.unpack_from(STRG_HEADER_STRUCT, self.strg, 0, getptr=True)
-		if hdata[0] != b'STRG':
-			error('Invalid STRG magic: %s' % byterepr(hdata[0]), 301)
-		#strglen = hdata[1]
-		i = 2
-		while i < 6:
-			id, offset = hdata[i:i + 2]
-			i += 2
-			if id == 0x2400:
-				fntoffset = offset + 8
-			elif id == 0x2401:
-				tbloffset = offset + 8
-		strcount = self.unpack_from('I', self.strg, fntoffset)[0]
-		strings = []
-		for i in range(0, strcount):
-			id, nameoffset, namelen = self.unpack_from('3I', self.strg, fntoffset + (i * 12) + 4)
-			name = self.strg[fntoffset + nameoffset:fntoffset + nameoffset + namelen].rstrip(b'\x00').decode('utf-8')
-			strings.append(name)
-		self.strings = strings
-
-	def readINFO(self):
-		hdata, ptr = self.unpack_from(INFO_HEADER_STRUCT, self.info, 0, getptr=True)
-		if hdata[0] != b'INFO':
-			error('Invalid INFO magic: %s' % byterepr(hdata[0]), 301)
-		infolen = hdata[1]
-		infooffsets = [None] * 7
-		i = 2
-		while i < 16:
-			id, offset = hdata[i:i + 2]
-			i += 2
-			if id == 0x2100:
-				infooffsets[0] = offset + 8
-			elif id == 0x2101:
-				infooffsets[1] = offset + 8
-			elif id == 0x2102:
-				infooffsets[2] = offset + 8
-			elif id == 0x2103:
-				infooffsets[3] = offset + 8
-			elif id == 0x2104:
-				infooffsets[4] = offset + 8
-			elif id == 0x2105:
-				infooffsets[5] = offset + 8
-			elif id == 0x2106:
-				infooffsets[6] = offset + 8
-			else:
-				error('Unsupported INFO section %x' % id, 303)
-		#unknown = hdata[16]
-		#unknown = hdata[18]
-		infosecs = [None] * 7
-		for i in range(0, len(infooffsets)):
-			startoffset = infooffsets[i]
-			try:
-				endoffset = getsup(infooffsets, startoffset)
-			except IndexError:
-				endoffset = infolen
-			infosecs[i] = self.info[startoffset: endoffset]
-		self.readFileTable(infosecs[6])
-		self.readBankTable(infosecs[1])
-		self.readAudioTable(infosecs[0])
-		self.readPlayerTable(infosecs[2])
-		self.readWARTable(infosecs[3])
-		self.readSetTable(infosecs[4])
-		self.readGroupTable(infosecs[5])
-
-	def readINFOsec(self, data):
-		entrycount, otbl, entrydata = self.unpack(INFO_SECTION_STRUCT, data)
-		tbllen = len(data) - len(entrydata) - 4
-		offsets = [el[1] - tbllen - 4 for el in otbl]
-		return entrycount, offsets, entrydata
-
-	def readAudioTable(self, data):
-		entrycount, offsets, entrydata = self.readINFOsec(data)
-		self.audios = [AudioNode() for i in range(0, entrycount)]
-		for i, offset in enumerate(offsets):
-			fileidx, ptr = self.uint32(entrydata, offset)
-			type, ptr = self.uint32(entrydata, ptr + 8)
-			stridx, ptr = self.uint32(entrydata, ptr + 8)
-			self.audios[i].fileid = fileidx
-			self.files[fileidx].name = self.strings[stridx]
-			self.audios[i].name = self.strings[stridx]
-			self.files[fileidx].fileid = i
-			if type == 0x2201:  #external stream
-				self.files[fileidx].type = self.audios[i].type = 'bcstm'
-			elif type == 0x2202:  #CWSD
-				self.files[fileidx].type = self.audios[i].type = 'bcwsd'
-			elif type == 0x2203:
-				self.files[fileidx].type = self.audios[i].type = 'bcseq'
-				self.audios[i].bank, ptr = self.uint16(entrydata, ptr + 0x44)
-
-	def readBankTable(self, data):
-		entrycount, offsets, entrydata = self.readINFOsec(data)
-		banktable = [self.unpack_from(BANK_ENTRY_STRUCT, entrydata, offset) for offset in offsets]
-		self.banks = [BankNode() for i in range(0, entrycount)]
-		for i, entry in enumerate(banktable):
-			fileidx = entry[0]
-			self.files[fileidx].type = 'bcbnk'
-			self.files[fileidx].bank = i
-			self.banks[i].fileid = fileidx
-			stridx = entry[4]
-			self.banks[i].name = self.strings[stridx]
-			self.files[fileidx].name = self.strings[stridx]
-
-	def readPlayerTable(self, data):
-		entrycount, offsets, entrydata = self.readINFOsec(data)
-		self.playertable = [self.unpack_from(PLAYER_ENTRY_STRUCT, entrydata, offset) for offset in offsets]
-
-	def readWARTable(self, data):
-		entrycount, offsets, entrydata = self.readINFOsec(data)
-		self.wars = [WARNode() for i in range(0, entrycount)]
-		for i, offset in enumerate(offsets):
-			fileidx, ptr = self.uint32(entrydata, offset)
-			self.wars[i].fileid = fileidx
-			self.files[fileidx].war = i
-			self.files[fileidx].type = 'bcwar'
-			entrylen = (offsets[i + 1] - offset) if i != len(offsets) - 1 else (len(entrydata) - offset)
-			if entrylen > 0x0c:
-				self.wars[i].name = self.uint32(entrydata, ptr)
-			else:
-				self.wars[i].name = 'WAR_%08x' % i
-			self.files[fileidx].name = self.wars[i].name
-
-	def readSetTable(self, data):
-		entrycount, offsets, entrydata = self.readINFOsec(data)
-		self.sets = [SetNode() for i in range(0, entrycount)]
-		for i, offset in enumerate(offsets):
-			hdr, ptr = self.uint32(entrydata, offset)
-			if hdr == 0xffffffff:
-				continue
-			stridx, ptr = self.uint32(entrydata, ptr + 0x18)
-			self.sets[i].name = self.strings[stridx]
-			fileidx, ptr = self.uint32(entrydata, ptr + 4)
-			self.sets[i].fileid = fileidx
-			self.files[fileidx].set = i
-
-	def readGroupTable(self, data):
-		entrycount, offsets, entrydata = self.readINFOsec(data)
-		grouptable = [self.unpack_from(GROUP_ENTRY_STRUCT, entrydata, offset) for offset in offsets]
-		self.groups = [GroupNode() for i in range(0, entrycount)]
-		for i, entry in enumerate(grouptable):
-			fileidx, unknown, stridx = entry
-			self.groups[i].fileid = fileidx
-			self.groups[i].name = self.strings[stridx]
-			self.files[fileidx].group = i
-			self.files[fileidx].name = self.strings[stridx]
-			self.files[fileidx].type = 'bcgrp'
-
-	def readFileTable(self, data):
-		entrycount, offsets, entrydata = self.readINFOsec(data)
-		self.files = [FileNode() for i in range(0, entrycount)]
-		for i, offset in enumerate(offsets):
-			id, ptr = self.uint32(entrydata, offset)
-			if id == 0x220d:  #external file
-				nameloc, ptr = self.uint32(entrydata, ptr)
-				self.files[i].filename, ptr = self.string(entrydata, offset + nameloc)
-			elif id == 0x220c:
-				self.files[i].internal = True
-				dataloc, ptr = self.uint32(entrydata, ptr)
-				if dataloc == 0xffffffff:
-					continue
-				type, ptr = self.uint32(entrydata, dataloc + offset)
-				self.files[i].offset, ptr = self.uint32(entrydata, ptr)
-				self.files[i].offset += 8
-				self.files[i].size, ptr = self.uint32(entrydata, ptr)
+				error('Invalid section ID 0x%04x' % id, 303)
 
 	def extract(self):
-		linked = []
-		for i, file in enumerate(self.files):
-			if file.internal:
-				if file.name == '':
-					print('file %x has no name' % i)
-					continue
-				elif file.type == 'Undefined':
-					print('file %x type is not defined' % i)
-				filename = '%s.%s' % (file.name, file.type)
-				filedata = self.file[file.offset: file.offset + file.size]
-				folder = self.get_folder(file.type)
-				bwrite(filedata, folder + filename)
-			else:
-				if file.filename is not None:
-					linked.append(file.filename)
-				else:
-					print('external file %x has no file name' % i)
-					continue
-		write('\n'.join(linked), self.outdir + 'LinkedExternalStreams.txt')
-		f, b, w, a = self.files, self.banks, self.wars, self.audios
-
-	def get_folder(self, type):
-		if type == 'bcseq':
-			return self.seqdir
-		elif type == 'bcwsd':
-			return self.wsddir
-		elif type == 'bcwar':
-			return self.wardir
-		elif type == 'bcbnk':
-			return self.bankdir
-		elif type == 'bcgrp':
-			return self.groupdir
-		else:
-			return self.outdir
+		self.file.extract(self.data)