+++ Changelog.md
@@ -259,3 +259,9 @@
 *	v1.20.48
 	*	Fixed the bug with the rotation on deswizzling of BFLIMs
 	*	Fixed a bug with absolute path in pack.GARC
+
+**v1.21**
+
+*	v1.21.48
+	*	Now, when packing without the -o option, 3DSkit will determin an output name from the input name and the format
+	*	Fixed multiple bugs in pack.BFLIM